+++ src/memory_migrate_plugin/adapters/__init__.py
@@ -1,6 +1,8 @@
 from .base import BaseAdapter
+from .claude_project import ClaudeProjectAdapter
 from .cline_memory_bank import ClineMemoryBankAdapter
 from .codex_memories import CodexMemoriesAdapter
+from .cursor_rules import CursorRulesAdapter
 from .generic_json import GenericJsonAdapter
 from .markdown_bundle import MarkdownBundleAdapter
 
@@ -10,4 +12,6 @@ __all__ = [
     "MarkdownBundleAdapter",
     "CodexMemoriesAdapter",
     "ClineMemoryBankAdapter",
+    "CursorRulesAdapter",
+    "ClaudeProjectAdapter",
 ]